Feature/ab#123456

+++ UserService/src/main/java/com/user/service/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
     User saveUser(User user);
     List<User> getAllUserDetails();
     User getUser(String userId);
-    int deleteUser(String userId);
+    long deleteUser(String userId);
     int updateUser(User user);
 
 }
